Decision tree bot integrated and builds
